==> all-of-us-notebooks/EBV_DNA_Quantification/03_Immunosuppressive_drug_exposure.ipynb <==
# Immunosuppressive drug exposure ~ EBV DNAemia

This notebook queries for individuals with immunosuppressive drug exposure and tests for associations with EBV DNAemia.

## Query drug exposures

The following RxNorm codes were obtained by querying https://athena.ohdsi.org/search-terms/terms?domain=Drug with the list of UKB immunosuppressive drug names (ED Table 2).

# RxNorm or other vocabs mapping to immmunosuppressive drugs
drug_concept_ids = [3211824, 19055344, 1551099, 903963, 1301267, 
                    1390051, 1310317, 955632, 35805577, 21201447, 
                    1350504, 1436650, 21256782, 1305058, 19014553, 
                    19014878, 42981494, 1377141, 21184231, 1351779, 
                    19011002, 1436678, 21200327, 19015514, 19014081, 
                    1366310, 21278185, 19033280, 1356461, 920458,
                    19013667, 19112543, 1518254, 19088949, 975125, 
                    1550557, 21293281, 21015053, 19015030, 19014302, 
                    19043341, 1506270, 21017639, 42799376, 1349025, 
                    1603790, 44022143, 19010400, 1603883, 19027526, 
                    1380068, 1379969, 21271639, 1525866, 45951973, 
                    45662921, 45699007, 3968922, 44348440, 1395557, 
                    950637, 1344381, 21182715, 1343039, 19010746,
                    1348265, 21275952, 713196, 19068900, 21259209,
                    1342346, 19086888, 1315946, 21219597, 42989336, 
                    722424, 19094015, 1368823, 3208281, 1350504, 1366310, 
                    21238100, 1341149, 1101898, 21258786, 21258787,
                    21219205, 3149129, 21237755, 3171040, 1398399, 
                    21275101, 950664, 3198126, 3187692, 751889, 
                    1337620, 21182376, 1507705, 19034726, 1604648,
                    1797155, 3206940, 21258075, 1114375, 21182275,
                    1714165, 3157849, 1304107, 1389888, 1389889, 
                    1119119, 3157931, 21275851, 19129200, 19129196,
                    21182684, 1304044, 3153286, 1603664, 21308962,
                    36123143, 21201789, 3181959, 21257518, 936429, 
                    2016567, 42981494, 46305620
]

# Convert list to a comma-separated string
drug_concept_ids_str = ', '.join(map(str, drug_concept_ids))

query="""SELECT person_id,
                drug_concept_id,
                drug_source_value,
                drug_source_concept_id,
                drug_exposure_start_date,
                drug_exposure_end_date
         FROM `"""+CDR_version+""".drug_exposure`
         WHERE drug_concept_id IN (3211824, 19055344, 1551099, 903963, 1301267, 1390051, 1310317, 955632, 35805577, 21201447, 1350504, 1436650, 21256782, 1305058, 19014553, 19014878, 42981494, 1377141, 21184231, 1351779, 19011002, 1436678, 21200327, 19015514, 19014081, 1366310, 21278185, 19033280, 1356461, 920458, 19013667, 19112543, 1518254, 19088949, 975125, 1550557, 21293281, 21015053, 19015030, 19014302, 19043341, 1506270, 21017639, 42799376, 1349025, 1603790, 44022143, 19010400, 1603883, 19027526, 1380068, 1379969, 21271639, 1525866, 45951973, 45662921, 45699007, 3968922, 44348440, 1395557, 950637, 1344381, 21182715, 1343039, 19010746, 1348265, 21275952, 713196, 19068900, 21259209, 1342346, 19086888, 1315946, 21219597, 42989336, 722424, 19094015, 1368823, 3208281, 1350504, 1366310, 21238100, 1341149, 1101898, 21258786, 21258787, 21219205, 3149129, 21237755, 3171040, 1398399, 21275101, 950664, 3198126, 3187692, 751889, 1337620, 21182376, 1507705, 19034726, 1604648, 1797155, 3206940, 21258075, 1114375, 21182275, 1714165, 3157849, 1304107, 1389888, 1389889, 1119119, 3157931, 21275851, 19129200, 19129196, 21182684, 1304044, 3153286, 1603664, 21308962, 36123143, 21201789, 3181959, 21257518, 936429, 2016567, 42981494, 46305620)"""
drugs=pd.read_gbq(query, dialect="standard")

drugs.to_csv("immunosuppressive_drugs_df.csv", sep=',')
